refactored as per feedback
