+++ src/components/BlockInserter.tsx
@@ -1,21 +1,20 @@
-import { useState } from 'react';
 import { BlockClient, BlockType } from '../../api/BlockClient';
 import { blockLibrary } from './blocks/BlockLibrary';
-import { Button, Flex, HStack } from '@chakra-ui/react';
+import { Button, Flex } from '@chakra-ui/react';
 
 interface BlockInserterProps {
-  position: number;
+  // position: number;
   quizId: string;
   stepId: string;
 }
 
-export const BlockInserter = ({ position, stepId, quizId }: BlockInserterProps) => {
+export const BlockInserter = ({ stepId, quizId }: BlockInserterProps) => {
   const addBlock = (blockType: BlockType) => {
     const newBlock = {
       quizId,
       stepId,
       type: blockType,
-      position: position,
+      // position: position,
       data: blockLibrary[blockType].factory(),
     };
 
@@ -23,13 +22,14 @@ export const BlockInserter = ({ position, stepId, quizId }: BlockInserterProps) 
   };
 
   return (
-    <Flex justify="space-around" p={2}>
-      <Button aria-label="select input block" onClick={() => addBlock(BlockType.INPUT)}>
-        {blockLibrary[BlockType.INPUT].inserterOptions.label}
-      </Button>
-      <Button aria-label="select textarea block" onClick={() => addBlock(BlockType.TEXTAREA)}>
-        {blockLibrary[BlockType.TEXTAREA].inserterOptions.label}
-      </Button>
+    <Flex justify="space-around" p={2} gap={2}>
+      {Object.keys(blockLibrary).map((block) => {
+        return (
+          <Button aria-label={`select ${block} block`} onClick={() => addBlock(block as BlockType)}>
+            {blockLibrary[block as BlockType].inserterOptions.label}
+          </Button>
+        );
+      })}
     </Flex>
   );
 };
